moved Sensors-POC into extra activity
