+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,14 @@
     <uses-permission android:name="android.permission.BLUETOOTH_ADMIN" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.CHANGE_WIFI_STATE" />
-
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
-    <uses-feature android:required="true" android:name="android.hardware.sensor.accelerometer"/>
-    <uses-feature android:required="true" android:name="android.hardware.sensor.light"/>
 
+    <uses-feature
+        android:name="android.hardware.sensor.accelerometer"
+        android:required="true" />
+    <uses-feature
+        android:name="android.hardware.sensor.light"
+        android:required="true" />
     <uses-feature
         android:name="android.hardware.sensor.proximity"
         android:required="true" />
@@ -29,10 +32,11 @@
         android:screenOrientation="landscape"
         android:supportsRtl="true"
         android:theme="@style/AppTheme.NoActionBar">
+        <activity android:name=".sensorspoc.SensorsPOC"></activity>
         <activity
             android:name=".GameEndActivity"
             android:label="@string/title_activity_game_end_"
-            android:theme="@style/AppTheme.NoActionBar.NoActionBar"></activity>
+            android:theme="@style/AppTheme.NoActionBar.NoActionBar" />
         <activity android:name=".lobby.ClientLobby" />
         <activity android:name=".lobby.GameCreate" />
         <activity
@@ -85,10 +89,10 @@
         </activity>
         <activity
             android:name=".lobby.GameJoin"
-            android:theme="@style/AppTheme.ActionBar"></activity>
+            android:theme="@style/AppTheme.ActionBar" />
         <activity
             android:name=".lobby.ServerLobby"
-            android:parentActivityName=".lobby.GameJoin"></activity>
+            android:parentActivityName=".lobby.GameJoin" />
     </application>
 
 </manifest>
