+++ app/src/main/java/com/example/scotlandyard/GameEndActivity.java
@@ -28,8 +28,6 @@ public class GameEndActivity extends AppCompatActivity {
         setContentView(R.layout.activity_game_end_);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-
-        sensors();
         initComp();
 
     }
@@ -49,75 +47,5 @@ public class GameEndActivity extends AppCompatActivity {
     public void openStart(View view) {
         startActivity(new Intent(GameEndActivity.this, MainActivity.class));
     }
-
-    private void sensors() {
-        SensorManager sm = (SensorManager) getSystemService(Context.SENSOR_SERVICE);
-        sm.registerListener(sensorListenerShaking, sm.getDefaultSensor(Sensor.TYPE_ACCELEROMETER), SensorManager.SENSOR_DELAY_NORMAL);
-        acelVal = SensorManager.GRAVITY_EARTH;
-        acelLast = SensorManager.GRAVITY_EARTH;
-        shake = 0.00f;
-
-        sm.registerListener(sensorListenerLight, sm.getDefaultSensor(Sensor.TYPE_LIGHT), SensorManager.SENSOR_DELAY_NORMAL);
-
-        sm.registerListener(sensorListenerProximity, sm.getDefaultSensor(Sensor.TYPE_PROXIMITY), SensorManager.SENSOR_DELAY_NORMAL);
-
-    }
-
-    private final SensorEventListener sensorListenerShaking = new SensorEventListener() {
-        @Override
-        public void onSensorChanged(SensorEvent event) {
-            float x = event.values[0];
-            float y = event.values[1];
-            float z = event.values[2];
-
-            acelLast = acelVal;
-            acelVal = (float) Math.sqrt((double)(x*x + y*y + z*z));
-            float delta = acelVal - acelLast;
-            shake = shake + delta;
-            //System.out.println(shake);
-            if(shake > 7){
-                Toast toast = Toast.makeText(getApplicationContext(), "Shake it baby", Toast.LENGTH_LONG);
-                toast.show();
-            }
-        }
-
-        @Override
-        public void onAccuracyChanged(Sensor sensor, int accuracy) {
-
-        }
-    };
-
-    private final SensorEventListener sensorListenerLight = new SensorEventListener() {
-        @Override
-        public void onSensorChanged(SensorEvent event) {
-            float curlight = event.values[0];
-            //System.out.println(curlight);
-            if(curlight < 100){
-                Toast toast = Toast.makeText(getApplicationContext(), "Ligths Out", Toast.LENGTH_LONG);
-                toast.show();
-            }
-        }
-
-        @Override
-        public void onAccuracyChanged(Sensor sensor, int accuracy) {
-
-        }
-    };
-
-    private final SensorEventListener sensorListenerProximity = new SensorEventListener() {
-        @Override
-        public void onSensorChanged(SensorEvent event) {
-            float distance = event.values[0];
-            if(distance < 5){
-                Toast toast = Toast.makeText(getApplicationContext(), "Cant Touch this", Toast.LENGTH_LONG);
-                toast.show();
-            }
-        }
-
-        @Override
-        public void onAccuracyChanged(Sensor sensor, int accuracy) {
-
-        }
-    };
 
 }
